Fixed compilation issue on osx

+++ divi/src/rpcrawtransaction.cpp
@@ -134,7 +134,7 @@ void TxToJSONExpanded(const CTransaction& tx, const uint256 hashBlock, Object& e
         }
     }
 
-    entry.push_back(Pair("size", tx.GetSerializeSize(SER_NETWORK, CTransaction::CURRENT_VERSION)));
+    entry.push_back(Pair("size", static_cast<int64_t>(tx.GetSerializeSize(SER_NETWORK, CTransaction::CURRENT_VERSION))));
 
 }
